LayoutTests/fast/dom/Window/window-screen-properties-expected.txt

PASS: window.screen.pixelDepth is 24
PASS: window.screen.colorDepth is 24
